Organise files for PR
