+++ SimulaQron/config/Nodes.cfg
@@ -1,10 +1,5 @@
-Node0
-Node1
-Node2
-Node3
-Node4
-Node5
-Node6
-Node7
-Node8
-Node9
+Alice
+Bob
+Charlie
+David
+Eve
