+++ SimulaQron/config/appNodes.cfg
@@ -5,13 +5,8 @@
 #
 # [name], [hostname], [port number]
 #
-Node0, localhost, 8821
-Node1, localhost, 8822
-Node2, localhost, 8823
-Node3, localhost, 8824
-Node4, localhost, 8825
-Node5, localhost, 8826
-Node6, localhost, 8827
-Node7, localhost, 8828
-Node8, localhost, 8829
-Node9, localhost, 8830
+Alice, localhost, 8811
+Bob, localhost, 8812
+Charlie, localhost, 8813
+David, localhost, 8814
+Eve, localhost, 8815
