+++ SimulaQron/config/cqcNodes.cfg
@@ -5,13 +5,8 @@
 #
 # [name], [hostname], [port number]
 #
-Node0, localhost, 8811
-Node1, localhost, 8812
-Node2, localhost, 8813
-Node3, localhost, 8814
-Node4, localhost, 8815
-Node5, localhost, 8816
-Node6, localhost, 8817
-Node7, localhost, 8818
-Node8, localhost, 8819
-Node9, localhost, 8820
+Alice, localhost, 8806
+Bob, localhost, 8807
+Charlie, localhost, 8808
+David, localhost, 8809
+Eve, localhost, 8810
