+++ SimulaQron/config/settings.ini
@@ -2,7 +2,7 @@
 maxqubits_per_node = 20
 maxregisters_per_node = 1000
 waittime = 0.5
-loglevel = debug
+loglevel = warning
 backendhandler = simulaqron
 topology_file = 
 noisy_qubits = False
