+++ SimulaQron/config/virtualNodes.cfg
@@ -5,13 +5,8 @@
 #
 # [name], [hostname], [port number]
 #
-Node0, localhost, 8801
-Node1, localhost, 8802
-Node2, localhost, 8803
-Node3, localhost, 8804
-Node4, localhost, 8805
-Node5, localhost, 8806
-Node6, localhost, 8807
-Node7, localhost, 8808
-Node8, localhost, 8809
-Node9, localhost, 8810
+Alice, localhost, 8801
+Bob, localhost, 8802
+Charlie, localhost, 8803
+David, localhost, 8804
+Eve, localhost, 8805
