+++ SimulaQron/cqc/backend/cqcConfig.py
@@ -33,7 +33,7 @@
 
 # Sets how long a node waits for receiving a qubit.
 # Raises a CQC_ERR_TIMEOUT when times is up.
-CQC_CONF_RECV_TIMEOUT = 5000  # (x 100 ms)
+CQC_CONF_RECV_TIMEOUT = 100  # (x 100 ms)
 CQC_CONF_RECV_EPR_TIMEOUT = 100  # (x 100 ms)
 CQC_CONF_WAIT_TIME_RECV = 0.1  # (seconds)  sets the time in seconds to wait before each recheck when receiving qubits
