+++ SimulaQron/examples/cqc/pythonLib/corrRNG/run.sh
@@ -2,4 +2,3 @@
 
 python aliceTest.py &
 python bobTest.py &
-#python charlieTest.py &
