--- WstateTest/README.txt
@@ -1,3 +1,0 @@
-In this example Alice and Bob generate correlated randomness by sharing an EPR pair and measuring.
-
-Alice creates two qubits and puts these into an EPR pair by performing a Hadamard and CNOT operation. She sends one of the qubit to Bob and measures the one she keeps in the standard-basis. Bob receives the qubit that Alice sent and measures it. Both print their measurement outcomes.
