--- WstateTest/Wstate/run.sh
@@ -1,23 +1,0 @@
-#!/bin/sh
-
-#python aliceTest.py &
-python default_Wstate_transmitter.py Node0 &
-python default_Wstate_receiver.py Node1 &
-python default_Wstate_receiver.py Node2 &
-python default_Wstate_receiver.py Node3 &
-python default_Wstate_receiver.py Node4 &
-python default_Wstate_receiver.py Node5 &
-python default_Wstate_receiver.py Node6 &
-python default_Wstate_receiver.py Node7 &
-python default_Wstate_receiver.py Node8 &
-python default_Wstate_receiver.py Node9 &
-#python default_Wstate_receiver.py Node10 &
-#python default_Wstate_receiver.py Node11 &
-#python default_Wstate_receiver.py Node12 &
-#python default_Wstate_receiver.py Node13 &
-#python default_Wstate_receiver.py Node14 &
-
-#python default_Wstate_receiver.py Bob &
-#python default_Wstate_receiver.py Eve &
-#python default_Wstate_receiver.py Charlie &
-#python default_Wstate_receiver.py David &
