--- WstateTest/butterfly/node_S1.py
@@ -1,69 +1,0 @@
-#
-# Copyright (c) 2017, Stephanie Wehner and Axel Dahlberg
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-# 1. Redistributions of source code must retain the above copyright
-#    notice, this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright
-#    notice, this list of conditions and the following disclaimer in the
-#    documentation and/or other materials provided with the distribution.
-# 3. All advertising materials mentioning features or use of this software
-#    must display the following acknowledgement:
-#    This product includes software developed by Stephanie Wehner, QuTech.
-# 4. Neither the name of the QuTech organization nor the
-#    names of its contributors may be used to endorse or promote products
-#    derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDER ''AS IS'' AND ANY
-# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER BE LIABLE FOR ANY
-# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-
-from SimulaQron.cqc.pythonLib.cqc import *
-
-
-#####################################################################################################
-#
-# main
-#
-def main():
-
-	# Initialize the connection
-	with CQCConnection("S1") as S1:
-
-		# Make EPR-pairs with R1 and T2
-		q2=S1.createEPR("R1")
-		q0=S1.createEPR("T2")
-
-		# Make Bell measurement (step 1)
-		q0.cnot(q2)
-		m=q2.measure()
-
-		# Send corrections to R1 (including sender) (step 1)
-		msg="S1".encode('utf-8')+bytes([m])
-		S1.sendClassical("R1",msg)
-
-		# Receive correction from R1 (step 7)
-		m=S1.recvClassical()
-		if m==1:
-			q0.Z()
-
-		# Measure out
-		m=q0.measure()
-		to_print="0: Measurement outcome: {}".format(m)
-		print("|"+"-"*(len(to_print)+2)+"|")
-		print("| "+to_print+" |")
-		print("|"+"-"*(len(to_print)+2)+"|")
-
-
-##################################################################################################
-main()
