--- WstateTest/butterfly/run.sh
@@ -1,6 +1,0 @@
-python node_S1.py &
-python node_S2.py &
-python node_R1.py &
-python node_R2.py &
-python node_T1.py &
-python node_T2.py &
