--- WstateTest/extendGHZ/README.txt
@@ -1,3 +1,0 @@
-In this example Alice, Bob and Charlie will produce a GHZ state between them and measure their qubits.
-
-Alice and Bob creates an EPR pair between them and Alice measures her qubit in the standard-basis. Bob creates a fresh qubit which he entangles with the qubit which is part of the EPR pair with Alice. Bob sends one of his qubits to Charlie and both Bob and Charlie measures their qubits in the standard-basis.
