--- WstateTest/extendGHZ/run.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-python aliceTest.py &
-python bobTest.py &
-python charlieTest.py &
