--- WstateTest/hello/README.txt
@@ -1,1 +1,0 @@
-In this example Alice connects to the cqc-server, sends a hello message and prints the return message
