--- WstateTest/hello/run.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-python aliceTest.py &
