--- WstateTest/run.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-python aliceTest.py &
-python bobTest.py &
