--- WstateTest/teleport/README.txt
@@ -1,4 +1,0 @@
-In this example Alice teleports a qubit to Bob.
-
-Alice and Bob starts by creating a shared EPR pair. Alice creates a fresh qubit which she wants to teleport. In this case she puts it in the |+> state by performing a Hadamard operation. She performes the local teleportation operations and measures her qubits. She sends her measurement outcomes using a classical channel. Bob receives the measurement outcomes and performs corrections if needed. He finally measures his qubit in the standard basis and prints the measurement
-outcome.
