+++ examples/cqc/pythonLib/wstate/doNew.sh
