+++ examples/cqc/pythonLib/wstate/run.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+
+python default_Wstate_transmitter.py Node0 &
+python default_Wstate_receiver.py Node1 &
+python default_Wstate_receiver.py Node2 &
+python default_Wstate_receiver.py Node3 &
+python default_Wstate_receiver.py Node4 &
+python default_Wstate_receiver.py Node5 &
+python default_Wstate_receiver.py Node6 &
